+++ Classes/BTUpdatable.h
@@ -4,5 +4,5 @@
 #import <Foundation/Foundation.h>
 
 @protocol BTUpdatable
-- (void)update:(float)dt;
+- (void)update:(double)dt;
 @end
